k8s: add helper methods for extracting pods
