+++ internal/k8s/image.go
@@ -2,7 +2,6 @@ package k8s
 
 import (
 	"fmt"
-	"reflect"
 
 	"github.com/docker/distribution/reference"
 	"k8s.io/api/core/v1"
@@ -70,63 +69,18 @@ func InjectImageDigest(entity K8sEntity, injectRef reference.Named, policy v1.Pu
 	return entity, replaced, nil
 }
 
-// Get pointers to all the container specs in this object.
-func extractContainers(obj interface{}) ([]*v1.Container, error) {
-	cType := reflect.TypeOf(v1.Container{})
-	v := reflect.ValueOf(obj)
-	result := make([]*v1.Container, 0)
-
-	// Recursively iterate over the struct fields.
-	var extract func(v reflect.Value) error
-	extract = func(v reflect.Value) error {
-		switch v.Kind() {
-		case reflect.Ptr, reflect.Interface:
-			if v.IsNil() {
-				return nil
-			}
-			return extract(v.Elem())
-
-		case reflect.Struct:
-			if v.Type() == cType {
-				if !v.CanAddr() {
-					return fmt.Errorf("Error addressing: %v", v)
-				}
-				ptr, ok := v.Addr().Interface().(*v1.Container)
-				if !ok {
-					return fmt.Errorf("Error addressing: %v", v)
-				}
-				result = append(result, ptr)
-				return nil
-			}
-
-			for i := 0; i < v.NumField(); i++ {
-				field := v.Field(i)
-				err := extract(field)
-				if err != nil {
-					return err
-				}
-			}
-			return nil
-
-		case reflect.Slice:
-			for i := 0; i < v.Len(); i++ {
-				field := v.Index(i)
-				err := extract(field)
-				if err != nil {
-					return err
-				}
-			}
-			return nil
-
+func PodContainsRef(pod *v1.PodSpec, ref reference.Named) (bool, error) {
+	for _, container := range pod.Containers {
+		existingRef, err := reference.ParseNormalizedNamed(container.Image)
+		if err != nil {
+			return false, fmt.Errorf("PodContainsRef: %v", err)
 		}
-		return nil
-	}
 
-	err := extract(v)
-	if err != nil {
-		return nil, err
+		if existingRef.Name() == ref.Name() {
+			return true, nil
+		}
 	}
-	return result, nil
+	return false, nil
 }
 
 func ParseNamedTagged(s string) (reference.NamedTagged, error) {
